Get rid of periodicity of the noise
Whatever the period, at some point, when failing to fall asleep, I do
notice the patterns, even if the period is as long as a minute. Further
increase of the period will lead to too long initialization of the sound
buffer, so now two buffers will be regenerated every 2 seconds and fade
one into another, thus avoiding any repetition.